add a integration test instead
