+++ library/src/androidTest/java/org/xmtp/android/library/InstrumentedTest.kt
@@ -1,5 +1,6 @@
 package org.xmtp.android.library
 
+import android.util.Log
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.FlakyTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -220,5 +221,21 @@ class InstrumentedTest {
         val options = ClientOptions(api = ClientOptions.Api(env = XMTPEnvironment.LOCAL, isSecure = true))
         val keys = PrivateKeyBundleV1Builder.buildFromBundle(bytes)
         Client().buildFrom(bundle = keys, options = options)
+    }
+
+    @Test
+    fun testPerformance() {
+        val privateKeyData = listOf(0x08, 0x36, 0x20, 0x0f, 0xfa, 0xfa, 0x17, 0xa3, 0xcb, 0x8b, 0x54, 0xf2, 0x2d, 0x6a, 0xfa, 0x60, 0xb1, 0x3d, 0xa4, 0x87, 0x26, 0x54, 0x32, 0x41, 0xad, 0xc5, 0xc2, 0x50, 0xdb, 0xb0, 0xe0, 0xcd)
+            .map { it.toByte() }
+            .toByteArray()
+        // Use hardcoded privateKey
+        val privateKey = PrivateKeyBuilder.buildFromPrivateKeyData(privateKeyData)
+        val privateKeyBuilder = PrivateKeyBuilder(privateKey)
+        val options = ClientOptions(api = ClientOptions.Api(XMTPEnvironment.DEV, isSecure = true))
+        val randomClient = Client().create(account = privateKeyBuilder, options = options)
+        val start = Date().time
+        val conversations = randomClient.conversations.list()
+        val end = Date().time
+        Log.d("PERF", "Loaded ${conversations.size} conversations in ${end - start}ms")
     }
 }
