+++ library/src/main/java/org/xmtp/android/library/Conversations.kt
@@ -142,7 +142,10 @@ data class Conversations(
 
     fun list(): List<Conversation> {
         val newConversations = mutableListOf<Conversation>()
+        val start = Date().time
         val seenPeers = listIntroductionPeers()
+        val end = Date().time
+        Log.d("PERF", "Loaded ${seenPeers.size} listIntroductionPeers in ${end - start}ms")
         for ((peerAddress, sentAt) in seenPeers) {
             newConversations.add(
                 Conversation.V1(
@@ -154,7 +157,10 @@ data class Conversations(
                 )
             )
         }
+        val start2 = Date().time
         val invitations = listInvitations()
+        val end2 = Date().time
+        Log.d("PERF", "Loaded ${invitations.size} listInvitations in ${end2 - start2}ms")
         for (sealedInvitation in invitations) {
             try {
                 val unsealed = sealedInvitation.v1.getInvitation(viewer = client.keys)
